+++ BankingApp/src/main/java/com/rajtech/bankapp/service/AccountService.java
@@ -6,9 +6,9 @@ import com.rajtech.bankapp.dto.AccountDto;
 
 public interface AccountService {
 	AccountDto openAccount(AccountDto accountDto);
-	AccountDto getAccountByAccountNumber(Long accountNumber);
-	AccountDto deposit(Long accountNumber, double depositAmount);
-	AccountDto withdraw(Long accountNumber, double withdrawAmount);
+	AccountDto getAccountByAccountNumber(int accountNumber);
+	AccountDto deposit(int accountNumber, double depositAmount);
+	AccountDto withdraw(int accountNumber, double withdrawAmount);
 	List<AccountDto> getAllAccounts();
-	void deleteAccount(Long accountNumber);
+	void deleteAccount(int accountNumber);
 }
